Updating namespace
Updating the project namespace from com.thedesertmonk to com.avojak
