+++ src/it/simple-it/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.thedesertmonk.mojo.it</groupId>
+    <groupId>com.avojak.mojo.it</groupId>
     <artifactId>simple-it</artifactId>
     <version>1.0-SNAPSHOT</version>
