+++ src/main/resources/resources.properties
@@ -1,11 +1,11 @@
-com.thedesertmonk.mojo.aws.p2.maven.plugin.AWSP2Mojo.info.skippingExecution=Skipping execution
-com.thedesertmonk.mojo.aws.p2.maven.plugin.AWSP2Mojo.info.skippingSnapshot=Skipping deployment of SNAPSHOT version
-com.thedesertmonk.mojo.aws.p2.maven.plugin.AWSP2Mojo.info.uploadComplete=Upload complete: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.exception.BucketDoesNotExistException.message=Bucket [{0}] does not exist
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.debug.deleteExistingObject=Deleting existing object: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.debug.skippingEmptyDirectory=Skipping upload of empty directory: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.debug.uploadingFile=Uploading file: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.warn.directoryContentsNull=Directory contents null: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.warn.directoryNotAccessible=Directory is not accessible: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.warn.fileNotAccessible=File is not accessible: {}
-com.thedesertmonk.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.error.failedUploadRequestCreation=Failed to create upload request
+com.avojak.mojo.aws.p2.maven.plugin.AWSP2Mojo.info.skippingExecution=Skipping execution
+com.avojak.mojo.aws.p2.maven.plugin.AWSP2Mojo.info.skippingSnapshot=Skipping deployment of SNAPSHOT version
+com.avojak.mojo.aws.p2.maven.plugin.AWSP2Mojo.info.uploadComplete=Upload complete: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.exception.BucketDoesNotExistException.message=Bucket [{0}] does not exist
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.debug.deleteExistingObject=Deleting existing object: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.debug.skippingEmptyDirectory=Skipping upload of empty directory: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.debug.uploadingFile=Uploading file: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.warn.directoryContentsNull=Directory contents null: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.warn.directoryNotAccessible=Directory is not accessible: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.warn.fileNotAccessible=File is not accessible: {}
+com.avojak.mojo.aws.p2.maven.plugin.s3.repository.S3BucketRepository.error.failedUploadRequestCreation=Failed to create upload request
